# src/swagger/swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Item Simulator REST APIs",
    "description": "Chapter 3, 개인과제 - 아이템 시뮬레이터 "
  },
  "servers": [
    {
      "url": "/api/",
      "description": ""
    }
  ],
...
  "paths": {
...
    "/character": {
      "post": {
        "tags": [
          "Characters: 캐릭터관련"
        ],
        "summary": "캐릭터 생성 API",
        "description": "[인증] 신규 캐릭터를 생성한다.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addCharacterReq"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Characters: 캐릭터관련"
        ],
        "summary": "캐릭터 목록 조회 API",
        "description": "[인증] 유저의 캐릭터 목록을 조회한다.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/character/{characterId}": {
      "delete": {
        "tags": [
          "Characters: 캐릭터관련"
        ],
        "summary": "캐릭터 삭제 API",
        "description": "[인증] 특정 캐릭터를 삭제한다.",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      },
      "get": {
        "tags": [
          "Characters: 캐릭터관련"
        ],
        "summary": "캐릭터 상세 조회 API",
        "description": "[인증] 특정 캐릭터를 조회한다. / [비인증] 게임머니 미출력",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "addCharacterReq": {
        "type": "object",
        "properties": {
          "characterName": {
            "type": "string",
            "example": "노래하는오징어"
          }
        },
        "required": [
          "characterName"
        ],
        "xml": {
          "name": "addCharacterReq"
        }
      },
...
    }
  }
}